+++ src/app/app.module.ts
@@ -41,7 +41,7 @@ import { AppComponent } from './app.component';
 import { AppRoutingModule, GoogleAnalyticsEffects } from 'routes';
 
 // Modules
-import { ErrorModule } from '@modules/error.module';
+import { ErrorModule } from '@app/pages/error/error.module';
 
 // Interceptor
 import { AppHttpInterceptor } from './interceptors/http.interceptor';
@@ -86,7 +86,7 @@ export function getMetaReducers(
 		PrebootModule.withConfig({ appRoot: 'app-root' }),
 		StoreModule.forRoot(
 			{
-				applicationState: fromApplication.Applicationreducer,
+				applicationState: fromApplication.reducer,
 				notificationState: fromNotifications.reducer,
 				routerState: routerReducer,
 			},
@@ -110,7 +110,7 @@ export function getMetaReducers(
 			provide: STORAGE_KEYS,
 			useValue: [
 				'applicationState',
-				'dummyState',
+				'entitiesState',
 				'notifications',
 				'routerState',
 			],