+++ src/app/pages/home/home.module.ts
@@ -2,14 +2,15 @@ import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
+import { MarkdownModule } from 'ngx-markdown';
 
 import { getInitialState } from '@app/ssr/tokens';
-import { DummyService } from '@services/dummy/dummy.service';
-import { DummyEffects } from '@store/dummy/dummy.effects';
-import * as fromDummy from '@store/dummy/dummy.reducer';
+import { EntitiesService } from '@app/services/entities/entities.service';
+import { EntitiesEffects } from '@app/store/entities/entities.effects';
+import * as fromEntities from '@app/store/entities/entities.reducer';
 import { HomePageComponent } from 'pages';
 import { HomeRoutingModule } from 'routes';
-import { SharedModule } from './shared.module';
+import { SharedModule } from '../../modules/shared.module';
 
 /**
  * Marks an class as an NgModule so it could be configured
@@ -18,14 +19,15 @@ import { SharedModule } from './shared.module';
 	declarations: [HomePageComponent],
 	imports: [
 		CommonModule,
-		EffectsModule.forFeature([DummyEffects]),
-		StoreModule.forFeature('dummyState', fromDummy.Dummyreducer, {
+		EffectsModule.forFeature([EntitiesEffects]),
+		MarkdownModule.forRoot(),
+		StoreModule.forFeature('entitiesState', fromEntities.reducer, {
 			initialState: getInitialState,
 		}),
 		SharedModule,
 		HomeRoutingModule,
 	],
-	providers: [DummyService],
+	providers: [EntitiesService],
 	schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
 /* istanbul ignore next */