+++ src/app/store/application/application.selectors.ts
@@ -1,9 +1,8 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { ApplicationState } from './application.reducer';
 
-export const selectFeature = createFeatureSelector<ApplicationState>(
-	'applicationState',
-);
+export const selectFeature =
+	createFeatureSelector<ApplicationState>('applicationState');
 
 export const selectIsMakingRequest = createSelector(
 	selectFeature,