+++ src/app/store/entities/entities.interface.ts
@@ -1,0 +1,12 @@
+export interface EntityInterface {
+	guid: string;
+	country: string;
+	currency: string;
+	dateFormat: string;
+	deeperObject: {
+		nestedAfterMap: string;
+		nestedBeforeMap: string;
+	};
+	email: string;
+	markdown: string;
+}