+++ src/environments/environment.prod.ts
@@ -1,5 +1,5 @@
 /**
- * The enviroment Object
+ * The environment Object
  */
 export const environment = {
 	/**
@@ -14,15 +14,15 @@ export const environment = {
 	/**
 	 * apiUrl will return the API url for selected enviroment
 	 */
-	apiUrl: `https://localhost:4000/api/v3/`,
+	apiUrl: `https://localhost:4000/api/`,
 	/**
 	 * assetsRoot provides the start of the URI for assets
 	 */
-	assetsRoot: 'https://localhost:4000/assets',
+	assetsRoot: 'https://localhost:4000/assets/',
 	/**
 	 * notificationServer will return the notificationServer url for selected enviroment
 	 */
-	notificationServer: `https://localhost:4000/v1`,
+	notificationServer: `https://localhost:4000/v1/`,
 	/**
 	 * Boolean to declare if the environment is in production mode
 	 */