+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG
 
+## [5.1.0](https://github.com/rickvandermey/angular-starterkit/tree/5.1.0)
+
+### Breaking Changes
+
+-   **Lighthouse:** Lighthouse updated report path
+-   **Compodoc:** revert to 1.1.11 if this error pops up: `Error: [BABEL] unknown: Preset /* your preset */ requires a filename to be set when babel is called directly,`
+
+### Features
+
+-   **NGRX:** removed dummy store and created an Entity store with a setAll function
+-   **NGRX:** added example of map functionality
+-   **Angular:** Updated to Angular 12.1.x
+-   **Packages:** Updated all dependencies to the latest version of 2021-07-15
+
 ## [5.0.0](https://github.com/rickvandermey/angular-starterkit/tree/5.0.0)
 
 ### Breaking Changes